qmp: remove LibvirtShellMonitor
